backend/requirements.txt
# ConectaBoi ETL - Backend Requirements

# API Framework
fastapi==0.104.1
uvicorn==0.24.0

# Configurações
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Processamento de dados
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Banco de dados
supabase==2.0.4
postgrest==0.15.0

# Utilitários
python-multipart==0.0.6
aiofiles==23.2.1

# Logging e monitoramento
structlog==23.2.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
